requirement.txt
Django==1.11
django-multi-token==0.1.2
django-request-logging==0.5.5
djangorestframework==3.7.7
pytz==2017.3
